--- a/src/types/boolean.ts
+++ b/src/types/boolean.ts
@@ -9,7 +9,7 @@ export interface Boolean extends Runtype<boolean> {
  */
 export const Boolean = create<Boolean>(
   x => {
-    if (typeof x !== 'boolean') throw validationError(`Expected boolean but was ${typeof x}`);
+    if (typeof x !== 'boolean') throw validationError(`Expected a boolean, but was ${typeof x}`);
     return x;
   },
   { tag: 'boolean' },