--- a/src/types/partial.ts
+++ b/src/types/partial.ts
@@ -2,6 +2,7 @@ import { Runtype, Static, create, validationError } from '../runtype';
 import { Union } from '../index';
 import { Undefined } from './literal';
 import { hasKey } from '../util';
+import show from '../show';
 
 export interface Part<O extends { [_: string]: Runtype }>
   extends Runtype<{ [K in keyof O]?: Static<O[K]> }> {
@@ -15,16 +16,22 @@ export interface Part<O extends { [_: string]: Runtype }>
 export function Part<O extends { [_: string]: Runtype }>(fields: O) {
   return create<Part<O>>(
     x => {
-      if (x === null || x === undefined)
-        throw validationError(`Expected a defined non-null value but was ${typeof x}`);
+      if (x === null || x === undefined) {
+        const a = create<any>(x, { tag: 'partial', fields });
+        throw validationError(`Expected a ${show(a)} value but was ${x}`);
+      }
 
       // tslint:disable-next-line:forin
       for (const key in fields)
         if (hasKey(key, x)) {
+          let FieldType = Union(fields[key], Undefined);
           try {
-            Union(fields[key], Undefined).check(x[key]);
-          } catch ({ message }) {
-            throw validationError(`In key ${key}: ${message}`);
+            FieldType.check(x[key]);
+          } catch ({ key: nestedKey }) {
+            throw validationError(
+              `Expected ${show(FieldType.reflect)}, got ${JSON.stringify(x[key])}`,
+              nestedKey ? `${key}.${nestedKey}` : key,
+            );
           }
         }