--- a/src/types/tuple.ts
+++ b/src/types/tuple.ts
@@ -232,8 +232,16 @@ export function Tuple(...components: Runtype[]): any {
       }
 
       if (xs.length < components.length)
-        throw validationError(`Expected array of ${components.length} but was ${xs.length}`);
-      for (let i = 0; i < components.length; i++) components[i].check(xs[i]);
+        throw validationError(`Expected an array of ${components.length}, but was ${xs.length}`);
+
+      for (let i = 0; i < components.length; i++) {
+        try {
+          components[i].check(xs[i]);
+        } catch ({ message, key: nestedKey }) {
+          throw validationError(message, nestedKey ? `[${i}].${nestedKey}` : `[${i}]`);
+        }
+      }
+
       return x;
     },
     { tag: 'tuple', components },