Goal was to improve the legibility of error messages, with emphasis on nested complex types.

Some examples of errors

```
import { Array, String, Number, Boolean, Record } from 'runtypes';
const Flags = Array(Boolean);
Flags.check([true, true, 's']);
// { Error: Expected boolean but was string, key: '[2]' }

const Page = Record({
  title: String
});

const Person = Record({
  name: String,
  age: Number,
  likes: Array(Page)
});
const People = Array(Person);
People.check([{ name: 'Jack', likes: [] }]);
// { Error: Expected age to be number, key: '[0].age' }

People.check([{ name: 'Jack', age: 10, likes: [{ name: undefined, likes: [] }] }]);
// { Error: Expected title to be string, key: '[0].likes.[0].title' }
```

~*Note:* Before merging, want to add some error specific tests (currently none of the tests checks the errors).~