@henrinormak 2018-04-16
Also improved union types, which now report the expected value along with the value that was checked.

```
const Pet = Record({
  name: String
});

const PersonOrPet = Union(
  Person,
  Pet
);

PersonOrPet.check({ foo: 'bar' });
// Error: None of the alternatives { name: string; age: number; likes: { title: string; }[]; },{ name: string; } matched {"foo":"bar"}
```