@henrinormak 2018-04-16
Finally, also improved errors for Partial type (including if the field is itself a complex type, such as a union):

```
const PartialPerson = Partial({
  name: String,
  age: Union(Number, String),
  likes: Array(Page)
});

PartialPerson.check({ name: 'Jack', age: { value: 10 } });
// { Error: Expected (number | string) | undefined, got {"value":10} key: 'age' }
```