@henrinormak 2018-04-16
Did some benchmarking, sort of as expected, when there is a fault, the new code is slower, with correct codepaths, the change remains within the standard deviation of the old error handling code (albeit it is slower). I used a large array (10k items) and a small one (10 items) of random data that checked against the Record type. Similarly with dictionaries.

```
Current production code - 8 tests completed.

  Correct array (large)      x     1,241 ops/sec ±10.85% (76 runs sampled)
  Correct dictionary (large) x       440 ops/sec ±7.77% (78 runs sampled)
  Faulty array (large)       x     1,374 ops/sec ±9.20% (83 runs sampled)
  Faulty dictionary (large)  x       445 ops/sec ±7.98% (81 runs sampled)
  Correct array (small)      x 1,412,369 ops/sec ±10.79% (80 runs sampled)
  Correct dictionary (small) x   401,691 ops/sec ±7.86% (50 runs sampled)
  Faulty array (small)       x    59,692 ops/sec ±7.24% (81 runs sampled)
  Faulty dictionary (small)  x    53,472 ops/sec ±6.55% (82 runs sampled)

With the new errors - 8 tests completed.

  Correct array (large)      x     1,465 ops/sec ±8.97% (83 runs sampled)
  Correct dictionary (large) x       432 ops/sec ±8.28% (78 runs sampled)
  Faulty array (large)       x     1,333 ops/sec ±10.12% (80 runs sampled)
  Faulty dictionary (large)  x       455 ops/sec ±4.93% (81 runs sampled)
  Correct array (small)      x 1,353,563 ops/sec ±9.58% (84 runs sampled)
  Correct dictionary (small) x   388,521 ops/sec ±9.10% (47 runs sampled)
  Faulty array (small)       x    41,303 ops/sec ±5.99% (81 runs sampled)
  Faulty dictionary (small)  x    35,289 ops/sec ±8.72% (76 runs sampled)
```